+++ swift/CareerCoreSwift/Sources/CareerCoreSmoke/main.swift
@@ -1,0 +1,5 @@
+import CareerCore
+import Foundation
+
+let output = try capabilitiesJson()
+FileHandle.standardOutput.write(Data(output.utf8))